1- Checked the currentUser role and check if the endpoint is allowed or forbidden.

+++ src/main/java/org/lms/course/CourseController.java
@@ -2,13 +2,16 @@ package org.lms.course;
 
 import jdk.jfr.ContentType;
 import org.apache.coyote.Request;
+import org.lms.authentication.interceptors.CurrentUser;
+import org.lms.user.User;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/course")
+@RequestMapping("/courses")
 public class CourseController {
 
     private CourseService courseService;
@@ -41,7 +44,10 @@ public class CourseController {
 
     // Add course
     @PostMapping("/create")
-    public ResponseEntity<Course> addCourse(@RequestBody CourseDTO course) throws Exception {
+    public ResponseEntity<Course> addCourse(@RequestBody CourseDTO course, @CurrentUser User currentUser) throws Exception {
+        if (currentUser.getRole() != User.Role.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         try {
             return ResponseEntity.ok().body(courseService.addCourse(course));
         }
@@ -52,7 +58,10 @@ public class CourseController {
 
     // Update course
     @PutMapping("/{id}")
-    public ResponseEntity<Course> updateCourse(@PathVariable("id") Integer id,@RequestBody CourseDTO course) throws Exception {
+    public ResponseEntity<Course> updateCourse(@PathVariable("id") Integer id,@RequestBody CourseDTO course, @CurrentUser User currentUser) throws Exception {
+        if (currentUser.getRole() != User.Role.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         try {
             return ResponseEntity.ok().body(courseService.updateCourse(id, course));
         }catch (Exception e) {
@@ -62,7 +71,11 @@ public class CourseController {
 
     // Delete course
     @DeleteMapping("/{id}")
-    public void deleteCourse(@PathVariable("id") Integer id) {
+    public ResponseEntity deleteCourse(@PathVariable("id") Integer id, @CurrentUser User currentUser) {
+        if (currentUser.getRole() != User.Role.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         courseService.deleteCourse(id);
+        return ResponseEntity.ok().build();
     }
 }
